check error condition

+++ utilities/piWrite.m
@@ -113,7 +113,9 @@ workingDir = thisR.get('output dir');
 % If we are using remote resources, remove leftover /local
 % files so they don't need to be rsynced
 if remoteResources
-    rmdir(workingDir, "s");
+    if isfolder(workingDir)
+        rmdir(workingDir, "s");
+    end
     mkdir(workingDir);
 % our traditional case:    
 elseif ~exist(workingDir,'dir')
